+++ api/lookup/controller.js
@@ -2,6 +2,7 @@ const log = require('log')('app:lookup');
 const Router = require('koa-router');
 const auth = require('../auth/auth');
 const service = require('./service');
+const _ = require('lodash');
 
 var router = new Router();
 
@@ -15,7 +16,7 @@ router.get('/api/lookup/application/enumerations', async (ctx, next) => {
     academicHonors: await service.lookupCodesByCodeType('academicHonors'),
     degreeTypes: await service.lookupCodesByCodeType('degreeTypes'),
     referenceTypes: await service.lookupCodesByCodeType('referenceTypes'),
-    securityClearances: await service.lookupCodesByCodeType('securityClearances'),
+    securityClearances: _.sortBy(await service.lookupCodesByCodeType('securityClearances'), 'code'),
   };
   
   ctx.body = result;
