+++ assets/js/backbone/apps/nav/templates/nav_template.html
@@ -25,7 +25,16 @@
                   </span>
                 </a>
               </li>
-              
+            <% } else if (name) { %>
+              <li class="usajobs-nav__menu-container usajobs-nav__account usajobs-nav--openopps__menu-container" data-state="is-closed">
+                <a class="usajobs-nav__section-link usajobs-nav--openopps__section-link" href="<%- usajobsURL %>/applicant/profile/openopportunities" title="Account" aria-haspopup="true" aria-controls="section-one">
+                  <span class="usajobs-nav--openopps__section">
+                    <span class="fas fa-user fa-fw" focusable="false"></span>
+                    <span class="usajobs-nav__first_name usajobs-nav--openopps__first-name"><%- name %></span>
+                    <span class="usajobs-nav__initials usajobs-nav--openopps__initials"><%- name %></span>
+                  </span>
+                </a>
+              </li>
             <% } else { %>
               <% if (login.signin !== false) { %>
               <li class="usajobs-nav__menu-container usajobs-nav__account">
@@ -44,14 +53,14 @@
                 </a>
               </li>
             <li class="usajobs-nav__menu-container usajobs-nav__menu-search desktop">
-              <a class="usajobs-nav__section-link usajobs-nav--openopps__section-link toggle-two" href="/search" title="Search Opportunities">
+              <a class="usajobs-nav__section-link usajobs-nav--openopps__section-link" href="/search" title="Search Opportunities">
                 <span class="usajobs-nav--openopps__section">
                   <span class="fas fa-search fa-fw" style="margin-top:0.5rem;" focusable="false"></span>Search
                 </span>
               </a>
             </li>
             <li class="usajobs-nav__menu-container usajobs-nav__menu-search usajobs-nav--openopps__menu-container mobile" data-state="is-closed">
-              <a class="usajobs-nav__section-link usajobs-nav--openopps__section-link menu-toggle toggle-two" href="#" title="Search Opportunities" aria-haspopup="true" aria-controls="section-two">
+              <a class="usajobs-nav__section-link usajobs-nav--openopps__section-link" href="/search" title="Search Opportunities" aria-haspopup="true" aria-controls="section-two">
                 <span class="usajobs-nav--openopps__section">
                   <span class="fas fa-search fa-fw" focusable="false"></span>Search
                 </span>
@@ -62,7 +71,7 @@
         <% } %>
       </div>
 
-      <% if (!accessForbidden && !user && loginGov) { %>
+      <% if (!accessForbidden && !welcome && !user && loginGov) { %>
       <div class="usajobs-nav-login-gov-banner" id="usajobs-nav-login-gov-banner" data-object="alert">
         <div class="usajobs-nav-login-gov-banner__body">
           <p class="usajobs-nav-login-gov-banner__text">
@@ -81,7 +90,7 @@
   </div>
 </nav>
 
-<% if(!accessForbidden) { %>
+<% if(!accessForbidden && !welcome) { %>
 <nav class="usajobs-openopps-secondary-nav">
   <div class="usajobs-openopps-secondary-nav__body">
     <% if (user) { %>
